Replace strapi with reading api
Deployed this microservice on fly.io so we can tear down strapi.
